ffend/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.5
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
google-generativeai==0.3.0
bcrypt==4.0.1
PyJWT==2.8.0
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0

# Watering Reminder System (Google Calendar)
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
APScheduler>=3.10.0
pytz>=2023.3
gspread>=5.12.0
